pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "etf-lp-am"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastparquet>=2025.12.0",
    "ipykernel>=7.2.0",
    "jupyter>=1.1.1",
    "jupytext>=1.19.1",
    "notebook>=7.5.3",
    "openpyxl>=3.1.5",
    "pandas>=3.0.0",
    "pyarrow>=23.0.0",
    "pydantic>=2.12.5",
    "pytest>=9.0.2",
    "python-dotenv>=1.2.1",
    "pyyaml>=6.0.3",
]

[tool.hatch.build.targets.wheel]
packages = ["utils"]

[tool.pytest.ini_options]
markers = [
    "integration: requires real CSV files in data/raw/chkxp/",
]
